FIX: small cleanup

+++ src/Traits/LogTrait.php
@@ -359,7 +359,9 @@ trait LogTrait
 
     protected function deleteLogFile()
     {
-        unlink($this->logFilePath());
+        if(file_exists($this->logFilePath())) {
+            unlink($this->logFilePath());
+        }
     }
